Applying comments: renamed configs
