+++ configs/xgboost/xgb_cpu_nvidia_config.json
