+++ datasets/load_datasets.py
@@ -21,12 +21,12 @@ import sys
 from pathlib import Path
 from typing import Callable, Dict
 
-from .loader_clf import (a_nine_a, airline, airline_ohe, bosch, codrnanorm,
-                         epsilon, fraud, gisette, higgs, higgs_one_m, ijcnn,
-                         klaverjas, santander, skin_segmentation)
-from .loader_mul import (connect, covertype, covtype, letters, mnist, msrank,
-                         plasticc, sensit)
-from .loader_reg import abalone, mortgage_first_q, year_prediction_msd
+from .loader_classification import (
+    a_nine_a, airline, airline_ohe, bosch, codrnanorm, epsilon, fraud, gisette,
+    higgs, higgs_one_m, ijcnn, klaverjas, santander, skin_segmentation)
+from .loader_multiclass import (connect, covertype, covtype, letters, mnist,
+                                msrank, plasticc, sensit)
+from .loader_regression import abalone, mortgage_first_q, year_prediction_msd
 
 dataset_loaders: Dict[str, Callable[[Path], bool]] = {
     "a9a": a_nine_a,
